# frontend/src/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FC Online Draft System",
    "description": "Realtime multi-team player draft system for Vietnamese FC Online pro tournament.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/tournaments": {
      "get": {
        "tags": [
          "Tournaments"
        ],
        "summary": "List Tournaments",
        "description": "List all tournaments (newest first).",
        "operationId": "list_tournaments_api_tournaments_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TournamentListItem"
                  },
                  "type": "array",
                  "title": "Response List Tournaments Api Tournaments Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tournaments"
        ],
        "summary": "Create Tournament",
        "description": "Create a new tournament with validated TournamentRules (ADMIN only).",
        "operationId": "create_tournament_api_tournaments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTournamentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{tournament_id}": {
      "get": {
        "tags": [
          "Tournaments"
        ],
        "summary": "Get Tournament",
        "description": "Get tournament details with teams.",
        "operationId": "get_tournament_api_tournaments__tournament_id__get",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tournament Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tournaments/{tournament_id}/draft/start": {
      "post": {
        "tags": [
          "Drafts"
        ],
        "summary": "Start Draft",
        "description": "Start a new draft session for a READY tournament (ADMIN only).",
        "operationId": "start_draft_api_tournaments__tournament_id__draft_start_post",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tournament Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{tournament_id}/draft": {
      "get": {
        "tags": [
          "Drafts"
        ],
        "summary": "Get Tournament Draft",
        "description": "Get the latest/active draft session details for a tournament.",
        "operationId": "get_tournament_draft_api_tournaments__tournament_id__draft_get",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tournament Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/drafts/{draft_id}": {
      "get": {
        "tags": [
          "Drafts"
        ],
        "summary": "Get Draft Detail",
        "description": "Get full draft details, current timer, and team status.",
        "operationId": "get_draft_detail_api_drafts__draft_id__get",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/drafts/{draft_id}/picks": {
      "get": {
        "tags": [
          "Drafts"
        ],
        "summary": "List Draft Picks",
        "description": "Get all picks in the draft session formatted for the draft board.",
        "operationId": "list_draft_picks_api_drafts__draft_id__picks_get",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DraftPickBoardItem"
                  },
                  "title": "Response List Draft Picks Api Drafts  Draft Id  Picks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Drafts"
        ],
        "summary": "Make Pick",
        "description": "Submit a single player card pick for the current team's turn.",
        "operationId": "make_pick_api_drafts__draft_id__picks_post",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftPickRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MakePickResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/drafts/{draft_id}/pause": {
      "post": {
        "tags": [
          "Drafts"
        ],
        "summary": "Pause Draft",
        "description": "Pause an active draft session (ADMIN only).",
        "operationId": "pause_draft_api_drafts__draft_id__pause_post",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartDraftResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/drafts/{draft_id}/resume": {
      "post": {
        "tags": [
          "Drafts"
        ],
        "summary": "Resume Draft",
        "description": "Resume a paused draft session (ADMIN only).",
        "operationId": "resume_draft_api_drafts__draft_id__resume_post",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartDraftResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/drafts/{draft_id}/cancel": {
      "post": {
        "tags": [
          "Drafts"
        ],
        "summary": "Cancel Draft",
        "description": "Cancel an active draft session (ADMIN only).",
        "operationId": "cancel_draft_api_drafts__draft_id__cancel_post",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartDraftResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTournamentRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "rules": {
            "$ref": "#/components/schemas/TournamentRulesSchema"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateTournamentRequest"
      },
...
      "DraftDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tournamentId": {
            "type": "string",
            "format": "uuid",
            "title": "Tournamentid"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "currentRound": {
            "type": "integer",
            "title": "Currentround"
          },
          "currentTurn": {
            "type": "integer",
            "title": "Currentturn"
          },
          "currentTeamId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currentteamid"
          },
          "turnStartedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turnstartedat"
          },
          "turnExpiresAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turnexpiresat"
          },
          "remainingMillis": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remainingmillis"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "serverTime": {
            "type": "string",
            "title": "Servertime"
          },
          "rulesSnapshot": {
            "additionalProperties": true,
            "type": "object",
            "title": "Rulessnapshot"
          },
          "teams": {
            "items": {
              "$ref": "#/components/schemas/TeamDraftStatus"
            },
            "type": "array",
            "title": "Teams"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tournamentId",
          "status",
          "currentRound",
          "currentTurn",
          "version",
          "serverTime",
          "rulesSnapshot",
          "teams"
        ],
        "title": "DraftDetailResponse"
      },
      "DraftPickBoardItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "round": {
            "type": "integer",
            "title": "Round"
          },
          "turnNumber": {
            "type": "integer",
            "title": "Turnnumber"
          },
          "salaryAtPick": {
            "type": "integer",
            "title": "Salaryatpick"
          },
          "pickedAt": {
            "type": "string",
            "title": "Pickedat"
          },
          "team": {
            "$ref": "#/components/schemas/PickTeamDetail"
          },
          "player": {
            "$ref": "#/components/schemas/PickPlayerDetail"
          }
        },
        "type": "object",
        "required": [
          "id",
          "round",
          "turnNumber",
          "salaryAtPick",
          "pickedAt",
          "team",
          "player"
        ],
        "title": "DraftPickBoardItem"
      },
      "DraftPickRequest": {
        "properties": {
          "playerSeasonId": {
            "type": "string",
            "format": "uuid",
            "title": "Playerseasonid"
          },
          "expectedVersion": {
            "type": "integer",
            "title": "Expectedversion"
          }
        },
        "type": "object",
        "required": [
          "playerSeasonId",
          "expectedVersion"
        ],
        "title": "DraftPickRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MakePickResponse": {
        "properties": {
          "pick": {
            "$ref": "#/components/schemas/PickResponseItem"
          },
          "nextState": {
            "$ref": "#/components/schemas/NextStateInfo"
          }
        },
        "type": "object",
        "required": [
          "pick",
          "nextState"
        ],
        "title": "MakePickResponse"
      },
...
      "TournamentListItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "rules": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TournamentRulesSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "teamCount": {
            "type": "integer",
            "title": "Teamcount"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "teamCount",
          "createdAt"
        ],
        "title": "TournamentListItem"
      },
      "TournamentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "rules": {
            "$ref": "#/components/schemas/TournamentRulesSchema"
          },
          "teams": {
            "items": {
              "$ref": "#/components/schemas/app__tournament__api__TeamSummary"
            },
            "type": "array",
            "title": "Teams"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "rules",
          "teams",
          "createdAt",
          "updatedAt"
        ],
        "title": "TournamentResponse"
      },
...
    }
  }
}